Rename src to lib since there's no babel anymore
