+++ package.json
@@ -8,9 +8,9 @@
   "scripts": {
     "release": "lerna publish",
     "release:force": "lerna publish --force-publish",
-    "test": "lerna exec --scope='*' --no-bail -- npm run test",
     "lint": "lerna exec --scope='*' --no-bail -- npm run lint",
     "lint:fix": "lerna exec --scope='*' --no-bail -- npm run lint:fix",
+    "test": "lerna exec --scope='*' --no-bail -- npm run test",
     "cover": "exit 0"
   },
   "devDependencies": {
