+++ packages/minim-api-description/lib/api-description.js
