+++ packages/minim-api-description/lib/elements/asset.js
