+++ packages/minim-api-description/lib/elements/auth-scheme.js
