+++ packages/minim-api-description/lib/elements/category.js
