+++ packages/minim-api-description/lib/elements/copy.js
