+++ packages/minim-api-description/lib/elements/resource.js
