+++ packages/minim-api-description/lib/elements/transition.js
